tests: Move fixture into conftest.py
This ensures that it is available in all test modules without importing
the unused function, see:
        https://docs.pytest.org/en/7.1.x/reference/fixtures.html#conftest-py-sharing-fixtures-across-multiple-files
