+++ tests/conftest.py
@@ -1,12 +1,15 @@
 from dataclasses import dataclass
 from os import environ
-from typing import Literal
+from typing import Iterator, Literal
+
+import pytest
 
 from nethsm import (
     DecryptMode,
     KeyMechanism,
     KeyType,
     LogLevel,
+    NetHSM,
     Role,
     UnattendedBootStatus,
 )
@@ -122,4 +125,21 @@ class Constants:
         ADMIN_USER,
     ]
 
-    # nitropy nethsm --host nethsmdemo.nitrokey.com --no-verify-tls info
+
+@pytest.fixture(scope="module")
+def nethsm() -> Iterator[NetHSM]:
+    """Start Docker container with Nethsm image and connect to Nethsm
+
+    This Pytest Fixture will run before the tests to provide the tests with
+    a nethsm instance via Docker container, also the first provision of the
+    NetHSM will be done in here"""
+
+    from utilities import connect, provision, start_nethsm
+
+    container = start_nethsm()
+
+    with connect(Constants.ADMIN_USER) as nethsm:
+        provision(nethsm)
+        yield nethsm
+
+    container.kill()
